+++ mods/enemies_reimagined/test.js
@@ -1,0 +1,8 @@
+// Just a demonstration of modularity
+// The mod loader will automatically load all .js files in the mods folder
+// You can use this to organize your code into multiple files
+// And you can also add completely new functions to the game
+// Just make sure to name them something unique so they don't conflict with other mods
+function NEW_FUNCTION_FROM_MOD() {
+  console.log("NEW_FUNCTION_FROM_MOD");
+}