+++ resources/compiledJS/skills & effects/stat_modifier.js
@@ -351,6 +351,22 @@ const traits = {
         },
     },
     /* End of starter trait choices */
+    // Secret traits
+    all_muscle: {
+        id: "all_muscle",
+        desc: "all_muscle_desc",
+        effects: {
+            strP: 15,
+            vitP: 15,
+            dexP: 15,
+            meleeDamageP: 30,
+            rangedDamageP: 30,
+            magicDamageP: -100,
+            mpMaxP: -100,
+            regenMpP: -100,
+            intP: -100,
+        },
+    },
 };
 class PermanentStatModifier {
     conditions;