+++ script/combat/damage.ts
@@ -2,7 +2,7 @@ function calculateDamage(attacker: characterObject, target: characterObject, abi
   // Initilize some values needed for the calculation
   const attackerStats = attacker.getStats();
   const targetResists = target.getResists();
-  const targetArmor = target.getArmor();
+  const targetArmor = target.getArmorReduction();
 
   // Roll for crit
   const critRolled = attackerStats.critChance >= helper.random(100, 0);
@@ -10,7 +10,8 @@ function calculateDamage(attacker: characterObject, target: characterObject, abi
   // Roll for evasion
   const hitChance = attacker.getHitchance().chance;
   const evasion = target.getHitchance().evasion;
-  const evade: boolean = evasion + helper.random(evasion * 0.5, evasion * -0.5) + 10 > hitChance + helper.random(hitChance * 0.3, hitChance * -0.6) + 20;
+  const evade: boolean =
+    evasion + helper.random(evasion * 0.5, evasion * -0.5) + 10 > hitChance + helper.random(hitChance * 0.3, hitChance * -0.6) + 20;
 
   // Create damage variable
   let damage: number = 0;
@@ -46,17 +47,24 @@ function calculateDamage(attacker: characterObject, target: characterObject, abi
 
     // Calculate bonus damage from stats
     let bonus: number = 0;
-    bonus += (damageValue * attackerStats[(attacker.weapon ? attacker.weapon.statBonus : attacker.firesProjectile ? "dex" : "str") ?? "str"]) / 50;
+    bonus +=
+      (damageValue * attackerStats[(attacker.weapon ? attacker.weapon.statBonus : attacker.firesProjectile ? "dex" : "str") ?? "str"]) / 50;
 
     // Calculate defense penetration
     let penetration = ability.resistance_penetration / 100;
 
     // Calculate defenses
-    let defense =
-      1 -
-      (targetArmor[damageCategories[damageType]] * 0.25 > 0 ? targetArmor[damageCategories[damageType]] * 0.25 * (1 - penetration) : targetArmor[damageCategories[damageType]]) /
-        100;
-    let resistance = 1 - (targetResists[damageType] > 0 ? targetResists[damageType] * (1 - penetration) : targetResists[damageType]) / 100;
+    let defense = 1;
+    const currentArmor = targetArmor[damageCategories[damageType]];
+    const penetrationMultiplier = 1 - penetration;
+    if (currentArmor > 0) {
+      const armorWithLoss = Math.min(currentArmor + penetrationMultiplier, 1);
+      defense = defense * armorWithLoss;
+    } else if (currentArmor) {
+      defense = defense * currentArmor;
+    }
+    let resistance =
+      1 - (targetResists[damageType] > 0 ? targetResists[damageType] * penetrationMultiplier : targetResists[damageType]) / 100;
 
     // Check for NaN to prevent breaking calculation
     if (isNaN(bonus)) bonus = 0;
@@ -66,7 +74,9 @@ function calculateDamage(attacker: characterObject, target: characterObject, abi
 
     // Calculate final damage
     let baseValue: number = damageValue + val + bonus;
-    let dmg = Math.floor(baseValue * mod * ability.damage_multiplier * (critRolled && !onlyRawDamage ? 1 + attackerStats.critDamage / 100 : 1) * defense);
+    let dmg = Math.floor(
+      baseValue * mod * ability.damage_multiplier * (critRolled && !onlyRawDamage ? 1 + attackerStats.critDamage / 100 : 1) * defense
+    );
     if (attackTypeDamageModifier > 0) dmg *= attackTypeDamageModifier;
     dmg = Math.floor(dmg * resistance);
     damage += dmg;